requirements.txt
bcrypt==4.1.2
passlib[bcrypt]
python-jose
sqlalchemy>=2.0
asyncmy
fastapi
uvicorn
dotenv
aiomysql
pydantic[email]
pydantic_settings
boto3
httpx
python-multipart
pandas